refactor: seperate recommendation components
